--- src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/pdf/helpers/Helpers.java
@@ -1,31 +1,0 @@
-package uk.gov.hmcts.reform.pip.publication.services.service.pdf.helpers;
-
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-
-/**
- * Class for static utility methods assisting with json->html->pdf issues.
- */
-public final class Helpers {
-
-    private Helpers() {
-        throw new UnsupportedOperationException();
-    }
-
-    public static String formatTimestampToBst(String timestamp) {
-        Instant unZonedDateTime = Instant.parse(timestamp);
-        ZoneId zone = ZoneId.of("Europe/London");
-        ZonedDateTime zonedDateTime = unZonedDateTime.atZone(zone);
-        DateTimeFormatter dtf;
-        dtf = DateTimeFormatter.ofPattern("dd MMMM yyyy 'at' HH:mm");
-        return dtf.format(zonedDateTime);
-    }
-
-    public static String formatLocalDateTimeToBst(LocalDateTime date) {
-        return date.format(
-            DateTimeFormatter.ofPattern("dd MMMM yyyy"));
-    }
-}
