+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/service/pdf/helpers/GeneralHelperTest.java
@@ -1,5 +1,5 @@
 
-package uk.gov.hmcts.reform.pip.publication.services.service.helpers;
+package uk.gov.hmcts.reform.pip.publication.services.service.pdf.helpers;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
